+++ tests/Router/RouterTest.php
@@ -93,7 +93,7 @@ class RouterTest extends TestCase
 
         self::$router = new Router($routes, [
             'basePath'    => '/',
-            'badRoute'    => 'E404',
+            //'badRoute'    => 'E404',
             'rewriteMode' => true,
         ]);
     }
@@ -203,8 +203,8 @@ class RouterTest extends TestCase
     public function routeProvider(): array
     {
         return [
-            ['/user', 'POST', ['E404Model', 'E404View', 'E404Controller', null, []], false], //test not allowed http method
-            ['/badroute', 'GET', ['E404Model', 'E404View', 'E404Controller', null, []], false], //test bad uri
+            //['/user', 'POST', ['E404Model', 'E404View', 'E404Controller', null, []], false], //test not allowed http method
+            //['/badroute', 'GET', ['E404Model', 'E404View', 'E404Controller', null, []], false], //test bad uri
             ['/user/5/enable', 'GET', ['UserModel', 'UserView', 'UserController', 'enable', ['id'=>'5']], true], //test param route
             ['/userOther/enable/5', 'GET', ['UserModel', 'UserView', 'UserController', 'enable', ['id'=>'5']], true], //test inverse param route
         ];
@@ -253,7 +253,7 @@ class RouterTest extends TestCase
     {
         $router = new Router(self::$routes, [
             'basePath'    => '/other_dir',
-            'badRoute'    => 'E404',
+            //'badRoute'    => 'E404',
             'rewriteMode' => true,
         ]);
 
@@ -415,7 +415,7 @@ class RouterTest extends TestCase
      *
      * @return void
      */
-    public function testValidateRouteWithNoBadRouteDeclared(): void
+    /*public function testValidateRouteWithNoBadRouteDeclared(): void
     {
         //using a worng bad route for overwrite previous setting
         $router = new Router(self::$routes, [
@@ -426,7 +426,7 @@ class RouterTest extends TestCase
         $this->assertFalse($router->validate('/badroute', 'GET'));
 
         $this->assertInstanceOf(NullRoute::class, $router->getRoute());
-    }
+    }*/
 
     /**
      * Test validate with rewrite mode off.
@@ -436,7 +436,7 @@ class RouterTest extends TestCase
     public function testValidateWithRewriteModeOff(): void
     {
         $router = new Router(self::$routes, [
-            'badRoute'    => 'E404',
+            //'badRoute'    => 'E404',
             'rewriteMode' => false,
         ]);
 
@@ -461,7 +461,7 @@ class RouterTest extends TestCase
     {
         $router = new Router(self::$routes, [
             'basePath'    => '/other_dir',
-            'badRoute'    => 'E404',
+            //'badRoute'    => 'E404',
             'rewriteMode' => false,
         ]);
 
@@ -547,7 +547,7 @@ class RouterTest extends TestCase
         ]));
 
         $router = new Router($restRoutes, [
-            'badRoute'    => 'E404',
+            //'badRoute'    => 'E404',
             'rewriteMode' => true,
         ]);
 
@@ -609,7 +609,7 @@ class RouterTest extends TestCase
      *
      * @return void
      */
-    public function testNotEqualRouteName(): void
+    /*public function testNotEqualRouteName(): void
     {
         $routes = (new RouteCollection([
             new Route([
@@ -638,7 +638,7 @@ class RouterTest extends TestCase
 
         $this->assertFalse($router->validate('/user/bad', 'GET'));
         $this->assertInstanceOf(NullRoute::class, $router->getRoute());
-    }
+    }*/
 
     /**
      * Route with param provider.
@@ -711,7 +711,7 @@ class RouterTest extends TestCase
 
         $router = new Router($routes, [
             'basePath'    => '/',
-            'badRoute'    => '404',
+            //'badRoute'    => '404',
             'rewriteMode' => true,
             'parseQueryStringOnRewriteModeOn' => true
         ]);